requirements.txt
asgiref==3.3.4
attrs==20.3.0
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
factory-boy==2.12.0
Faker==8.1.1
inflection==0.5.1
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.3
pytest-django==4.2.0
pytest-factoryboy==2.0.3
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
